test(logic): split orders work completion tests under test/orders
Move the former monolithic orders_application_work_completion_test.dart
into three focused files (build improvement, explore/roads, port/fort/rail),
each under the 400-line logic test gate and aligned with lib/src/orders.

Refs #2288.
