+++ packages/colonizethis_logic/test/orders/orders_application_work_completion_infrastructure_test.dart
@@ -1,0 +1,224 @@
+import 'package:colonizethis_test/test.dart';
+import 'package:colonizethis_data/colonizethis_data.dart';
+import 'package:colonizethis_logic/colonizethis_logic.dart';
+import 'package:colonizethis_models/colonizethis_models.dart';
+
+
+void main() {
+  group('applyBuildAndWorkOrders work completion (port, fort, rail)', () {
+    const ow = 'oldWorld';
+    const tileKey = 'oldWorld|P1|0|0';
+    const provinceId = 'oldWorld|P1';
+
+    // Non-empty orders so applyBuildAndWorkOrders does not return early (empty build list still counts).
+    Orders ordersToTriggerProcessWork() =>
+        Orders(buildUnitOrdersByPlayerId: {'p1': <BuildUnitOrder>[]});
+
+    test(
+      'build_port completion sets port and road level 4 when topology has sea',
+      () {
+        final topology = MapTopology(
+          nodes: const [
+            TopologyNode(
+              id: 'P1',
+              regionId: ow,
+              type: TopologyNodeType.province,
+            ),
+            TopologyNode(
+              id: 'sea1',
+              regionId: ow,
+              type: TopologyNodeType.seaZone,
+            ),
+          ],
+          edges: const [TopologyEdge(id1: 'P1', id2: 'sea1')],
+        );
+        final unit = Unit(
+          id: 'u1',
+          type: kUnitTypeEngineer,
+          ownerId: 'p1',
+          locationProvinceId: provinceId,
+          tileKey: tileKey,
+          status: UnitStatus.working,
+          currentWork: const CurrentWork(
+            workTarget: kWorkTargetBuildPort,
+            tileKey: tileKey,
+            totalTurns: 1,
+            remainingTurns: 1,
+          ),
+        );
+        final game = Game(
+          id: 'g',
+          worldState: WorldState(
+            turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 0),
+            oldWorld: RegionData(
+              provinces: [
+                Province(id: provinceId, regionId: ow, ownerId: 'p1'),
+              ],
+              units: [unit],
+            ),
+            newWorld: const RegionData(),
+          ),
+          players: const [Player(id: 'p1', displayName: 'P1', isHuman: true)],
+        );
+        final next = applyBuildAndWorkOrders(
+          game,
+          ordersToTriggerProcessWork(),
+          topology: topology,
+        );
+        expect(next.worldState.tileState.roadLevel(tileKey), 4);
+        expect(
+          next.worldState.portsByProvinceSeaboard.keys.any(
+            (k) => k.startsWith(provinceId),
+          ),
+          isTrue,
+        );
+      },
+    );
+
+    test('build_fort completion increases province fortLevel', () {
+      final unit = Unit(
+        id: 'u1',
+        type: kUnitTypeEngineer,
+        ownerId: 'p1',
+        locationProvinceId: provinceId,
+        tileKey: tileKey,
+        status: UnitStatus.working,
+        currentWork: const CurrentWork(
+          workTarget: kWorkTargetBuildFort,
+          tileKey: tileKey,
+          totalTurns: 1,
+          remainingTurns: 1,
+        ),
+      );
+      final game = Game(
+        id: 'g',
+        worldState: WorldState(
+          turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 0),
+          oldWorld: RegionData(
+            provinces: [
+              Province(
+                id: provinceId,
+                regionId: ow,
+                ownerId: 'p1',
+                fortLevel: 0,
+              ),
+            ],
+            units: [unit],
+          ),
+          newWorld: const RegionData(),
+        ),
+        players: const [Player(id: 'p1', displayName: 'P1', isHuman: true)],
+      );
+      final next = applyBuildAndWorkOrders(game, ordersToTriggerProcessWork());
+      expect(next.worldState.oldWorld.provinces.single.fortLevel, 1);
+    });
+
+    test('build_rail completion leaves road when tile has no road', () {
+      final tileState = TileMapState().setRoadLevel(tileKey, 0);
+      final unit = Unit(
+        id: 'u1',
+        type: kUnitTypeRailBuilder,
+        ownerId: 'p1',
+        locationProvinceId: provinceId,
+        tileKey: tileKey,
+        status: UnitStatus.working,
+        currentWork: const CurrentWork(
+          workTarget: kWorkTargetBuildRail,
+          tileKey: tileKey,
+          totalTurns: 1,
+          remainingTurns: 1,
+        ),
+      );
+      final railMap = TileMapResult(
+        width: 1,
+        height: 1,
+        grid: const [
+          ['P1'],
+        ],
+        terrainGrid: [
+          [TerrainType.plains],
+        ],
+      );
+      final game = Game(
+        id: 'g',
+        worldState: WorldState(
+          turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 0),
+          oldWorld: RegionData(
+            provinces: [Province(id: provinceId, regionId: ow, ownerId: 'p1')],
+            units: [unit],
+          ),
+          newWorld: const RegionData(),
+          tileState: tileState,
+        ),
+        players: [
+          Player(
+            id: 'p1',
+            displayName: 'P1',
+            isHuman: true,
+            techUnlocked: const {kTechIdEarlySteamEngine: true},
+          ),
+        ],
+      );
+      final next = applyBuildAndWorkOrders(
+        game,
+        ordersToTriggerProcessWork(),
+        tileMapByRegion: {ow: railMap},
+      );
+      expect(next.worldState.tileState.roadLevel(tileKey), 0);
+    });
+
+    test('build_rail completion sets road level to 4 when valid', () {
+      final tileState = TileMapState().setRoadLevel(tileKey, 1);
+      final unit = Unit(
+        id: 'u1',
+        type: kUnitTypeRailBuilder,
+        ownerId: 'p1',
+        locationProvinceId: provinceId,
+        tileKey: tileKey,
+        status: UnitStatus.working,
+        currentWork: const CurrentWork(
+          workTarget: kWorkTargetBuildRail,
+          tileKey: tileKey,
+          totalTurns: 1,
+          remainingTurns: 1,
+        ),
+      );
+      final railMap = TileMapResult(
+        width: 1,
+        height: 1,
+        grid: const [
+          ['P1'],
+        ],
+        terrainGrid: [
+          [TerrainType.plains],
+        ],
+      );
+      final game = Game(
+        id: 'g',
+        worldState: WorldState(
+          turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 0),
+          oldWorld: RegionData(
+            provinces: [Province(id: provinceId, regionId: ow, ownerId: 'p1')],
+            units: [unit],
+          ),
+          newWorld: const RegionData(),
+          tileState: tileState,
+        ),
+        players: [
+          Player(
+            id: 'p1',
+            displayName: 'P1',
+            isHuman: true,
+            techUnlocked: const {kTechIdEarlySteamEngine: true},
+          ),
+        ],
+      );
+      final next = applyBuildAndWorkOrders(
+        game,
+        ordersToTriggerProcessWork(),
+        tileMapByRegion: {ow: railMap},
+      );
+      expect(next.worldState.tileState.roadLevel(tileKey), 4);
+    });
+  });
+}
